requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.2
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
